% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Example_data}
\alias{Example_data}
\title{Simulated randomized experiment data with clustering I}
\format{
A data frame with 100 rows and 21 columns
\describe{
\item{id}{Bbservation ID}
\item{cl}{Assigned cluster number}
\item{cid}{Cluster-specific identifier for each observation within its cluster. It is assigned based on the cluster assignment \code{cl} and the \code{count} parameter.}
\item{X1...X3}{Covariates with each observation}
\item{prob}{Probability of treatment assignment}
\item{Y}{Observed outcomes}
\item{trt}{Treatment assignment for each observation}
\item{Yc0t1...Yc5t1}{Counterfactual outcome for the scenario where an observation belongs to cluster \code{c} and is assigned to treatment \code{t}.}
\item{Y0...Y5}{Counterfactual outcome for the scenario where an observation is assigned to a specific treatment}
}
}
\usage{
Example_data
}
\description{
A data set where each row corresponds to an individual and columns contain information on treatments,
covariates, probabilities of treatment assignment, and observed outcomes.
This simulation give more weights to X1 when determining the outcome and it is the approach used in simulations performed in manuscript.
}
\keyword{datasets}
