% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Example_data_non_clus}
\alias{Example_data_non_clus}
\title{Simulated randomized experiment data without clustering}
\format{
A data frame with 100 rows and 13 columns
\describe{
\item{id}{Bbservation ID}
\item{trt}{Treatment assigned for each observation}
\item{X1}{Covariate X1}
\item{X2}{Covariate X2}
\item{X3}{Covariate X3}
\item{prob}{Probability of treatment assignment}
\item{Y}{Observed outcomes}
\item{Y0}{Counterfactual outcome for the scenario where an observation is assigned to treatment 0}
\item{Y1}{Counterfactual outcome for the scenario where an observation is assigned to treatment 1}
\item{Y2}{Counterfactual outcome for the scenario where an observation is assigned to treatment 2}
\item{Y3}{Counterfactual outcome for the scenario where an observation is assigned to treatment 3}
\item{Y4}{Counterfactual outcome for the scenario where an observation is assigned to treatment 4}
\item{Y5}{Counterfactual outcome for the scenario where an observation is assigned to treatment 5}
}
}
\usage{
Example_data_non_clus
}
\description{
A data set where each row corresponds to an individual and columns contain information on treatments,
covariates, probabilities of treatment assignment, and observed outcomes
}
\keyword{datasets}
