% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle.R
\name{oracle}
\alias{oracle}
\title{Generate an oracle solution that selects the treatment with the highest response value for each individual in validation data}
\usage{
oracle(data, y, id, trt)
}
\arguments{
\item{data}{validation data}

\item{y}{a character string indicating the column name of outcomes.}

\item{id}{a vector of character strings indicating the column names of IDs}

\item{trt}{a vector of character strings indicating the column names of treatments}
}
\value{
original data alongside additional columns representing the oracle-selected treatment and response variable values. These additional columns are derived from the original data based on the treatment variable's highest associated response value for each individual.
}
\description{
Generate an oracle solution that selects the treatment with the highest response value for each individual in validation data
}
\examples{
\dontrun{
data(Example_data)
y <- "Y"; trt <- "trt"; id <- "id"

Example_trainest <- Example_data \%>\% slice_sample(n = floor(0.3 * nrow(Example_data)))
Example_valid <- Example_data \%>\% filter(!id \%in\% Example_trainest$id)

Example_oracle <- oracle(Example_valid, y, id, trt)
}

}
